@@ -1,41 +1,30 @@
-import "./styles.css";
+import logo from './logo.svg';
+import './App.css';
+import Clock from './components/Clock';
+import Date from './components/Date';
+import Heat from './components/Heat';
+import Settings from './components/Settings';
 
-export default function App() {
+function App() {
   return (
-    <div className="container mx-auto text-center">
-      <h1 className="text-3xl text-black-900 p-5 font-extrabold tracking-wide font-sans">
-        Hello User
-      </h1>
-      <h2 className="text-black-300 text-lg tracking-wide">
-        Enter Your City Name to Get Info about Weather
-      </h2>
-      <input
-        className="border-2 py-2 px-5 my-5 rounded-3xl w-8/12 hover:ring-2"
-        type="text"
-        placeholder="City Name"
-      />
-      <div className="information">
-        <div className="row flex py-5 ">
-          <div className="column bg-black-100 w-4/12 mx-2 border-2 p-5">
-            <h1> City Name : Udaipur </h1>
-            <p>
-              {" "}
-              Lorem ipsum dolor sit amet consectetur adipiscing elit, maecenas
-              ac non bibendum dictumst justo tempus ut, phasellus mattis vel eu
-              metus montes.{" "}
-            </p>
-          </div>
-          <div className="column bg-black-100 w-8/12 mx-2 border-2 p-5">
-            <h1> City Name : Udaipur </h1>
-            <p>
-              {" "}
-              Lorem ipsum dolor sit amet consectetur adipiscing elit, maecenas
-              ac non bibendum dictumst justo tempus ut, phasellus mattis vel eu
-              metus montes.{" "}
-            </p>
-          </div>
+    <div className="container">
+      <div className="app-container">
+        <div className="app-top-bar">
+          <button className="button button-small">
+            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-brightness-alt-high" viewBox="0 0 16 16">
+              <path d="M8 3a.5.5 0 0 1 .5.5v2a.5.5 0 0 1-1 0v-2A.5.5 0 0 1 8 3zm8 8a.5.5 0 0 1-.5.5h-2a.5.5 0 0 1 0-1h2a.5.5 0 0 1 .5.5zm-13.5.5a.5.5 0 0 0 0-1h-2a.5.5 0 0 0 0 1h2zm11.157-6.157a.5.5 0 0 1 0 .707l-1.414 1.414a.5.5 0 1 1-.707-.707l1.414-1.414a.5.5 0 0 1 .707 0zm-9.9 2.121a.5.5 0 0 0 .707-.707L3.05 5.343a.5.5 0 1 0-.707.707l1.414 1.414zM8 7a4 4 0 0 0-4 4 .5.5 0 0 0 .5.5h7a.5.5 0 0 0 .5-.5 4 4 0 0 0-4-4zm0 1a3 3 0 0 1 2.959 2.5H5.04A3 3 0 0 1 8 8z" />
+            </svg>
+          </button>
+          <h1 className="app-heading">Weather Stats</h1>
         </div>
+        <Date />
+        <Clock />
+        <Heat />
+        <Settings />
       </div>
+
     </div>
   );
 }
+
+export default App;
